+++ bench/hist/Main.hs
@@ -47,31 +47,19 @@
 {-# LANGUAGE DeriveAnyClass    #-}
 {-# LANGUAGE DerivingStrategies#-}
 {-# LANGUAGE TypeFamilies      #-}
+{-# OPTIONS -Wno-orphans #-}
 
-import Control.Applicative (Alternative (empty))
-import Control.Monad (when, forM, forM_, replicateM)
-import Data.Char (toLower)
 import Data.Foldable (find)
 import Data.Maybe (fromMaybe)
-import Data.Text (Text)
-import qualified Data.Text as T
-import Data.Yaml ((.!=), (.:?), FromJSON (..), ToJSON (..), Value (..), decodeFileThrow)
+import Data.Yaml (FromJSON (..), decodeFileThrow)
+import Development.Benchmark.Rules
 import Development.Shake
-import Development.Shake.Classes (Binary, Hashable, NFData)
-import Experiments.Types (getExampleName, exampleToOptions, Example(..))
-import GHC.Exts (IsList (..))
+import Experiments.Types (Example, exampleToOptions)
+import qualified Experiments.Types as E
 import GHC.Generics (Generic)
-import qualified Graphics.Rendering.Chart.Backend.Diagrams as E
-import Graphics.Rendering.Chart.Easy ((.=))
-import qualified Graphics.Rendering.Chart.Easy as E
 import Numeric.Natural (Natural)
 import System.Directory
-import System.FilePath
-import qualified Text.ParserCombinators.ReadP as P
-import Text.Read (Read (..), get, readMaybe, readP_to_Prec)
-import GHC.Stack (HasCallStack)
-import Data.List (transpose)
-import GHC.Records (HasField(getField))
+
 
 config :: FilePath
 config = "bench/config.yaml"
@@ -80,268 +68,23 @@ config = "bench/config.yaml"
 readConfigIO :: FilePath -> IO (Config BuildSystem)
 readConfigIO = decodeFileThrow
 
-newtype GetExample = GetExample String deriving newtype (Binary, Eq, Hashable, NFData, Show)
-newtype GetExamples = GetExamples () deriving newtype (Binary, Eq, Hashable, NFData, Show)
-newtype GetSamples = GetSamples () deriving newtype (Binary, Eq, Hashable, NFData, Show)
-newtype GetExperiments = GetExperiments () deriving newtype (Binary, Eq, Hashable, NFData, Show)
-newtype GetVersions = GetVersions () deriving newtype (Binary, Eq, Hashable, NFData, Show)
-newtype GetParent = GetParent Text deriving newtype (Binary, Eq, Hashable, NFData, Show)
-newtype GetCommitId = GetCommitId String deriving newtype (Binary, Eq, Hashable, NFData, Show)
-newtype GetBuildSystem = GetBuildSystem () deriving newtype (Binary, Eq, Hashable, NFData, Show)
-
+instance IsExample Example where getExampleName = E.getExampleName
 type instance RuleResult GetExample = Maybe Example
 type instance RuleResult GetExamples = [Example]
-type instance RuleResult GetSamples = Natural
-type instance RuleResult GetExperiments = [Unescaped String]
-type instance RuleResult GetVersions = [GitCommit]
-type instance RuleResult GetParent = Text
-type instance RuleResult GetCommitId = String
-type instance RuleResult GetBuildSystem = BuildSystem
 
 main :: IO ()
 main = shakeArgs shakeOptions {shakeChange = ChangeModtimeAndDigest} $ do
-  createBuildSystem readConfigIO $ \build resource -> do
+  createBuildSystem $ \resource -> do
+      configStatic <- liftIO $ readConfigIO config
+      let build = outputFolder configStatic
       buildRules build ghcideBuildRules
-      benchRules build resource ghcideBuildRules
+      benchRules build resource (MkBenchRules (benchGhcide $ samples configStatic) "ghcide")
       csvRules build
       svgRules build
-      action $ defaultAll build
+      action $ allTargets build
 
-
-ghcideBuildRules :: MkBuildRules BuildSystem Example
-ghcideBuildRules = MkBuildRules findGhcDefault buildGhcide benchGhcide
-
-data MkBuildRules buildSystem example = MkBuildRules
-  { findGhc            :: buildSystem -> FilePath -> IO FilePath
-  , buildProject       :: buildSystem -> FilePath -> String
-  , benchProject       :: buildSystem -> BenchProject example -> [String]
-  }
-
-data BenchProject example = BenchProject
-    { outcsv  :: FilePath
-    , ghcPath :: FilePath
-    , example :: example
-    , experiment :: Escaped String
-    , samples :: Natural
-    }
-
--- TODO Generalize
-createBuildSystem :: (FilePath -> IO (Config BuildSystem)) -> (String -> Resource -> Rules a) -> Rules a
-createBuildSystem readConfigFromFile userRules = do
-  readConfig <- newCache $ \fp -> need [fp] >> liftIO (readConfigFromFile fp)
-
-  _ <- addOracle $ \GetSamples {} -> getField @"samples" <$> readConfig config
-  _ <- addOracle $ \GetExperiments {} -> experiments <$> readConfig config
-  _ <- addOracle $ \GetVersions {} -> versions <$> readConfig config
-  _ <- addOracle $ \GetExamples{} -> examples <$> readConfig config
-  _ <- addOracle $ \(GetParent name) -> findPrev name . versions <$> readConfig config
-  _ <- addOracle $ \(GetExample name) -> find (\e -> getExampleName e == name) . examples <$> readConfig config
-  _ <- addOracle $ \GetBuildSystem {} -> buildTool <$> readConfig config
-
-  benchResource <- newResource "ghcide-bench" 1
-
-  configStatic <- liftIO $ readConfigIO config
-  let build = outputFolder configStatic
-  userRules build benchResource
-
-
-defaultAll :: FilePath -> Action ()
-defaultAll build = do
-    experiments <- askOracle $ GetExperiments ()
-    examples    <- askOracle $ GetExamples ()
-    versions    <- askOracle $ GetVersions ()
-    need $
-      [build </> getExampleName e </> "results.csv" | e <- examples ] ++
-      [build </> "results.csv"]
-        ++ [ build </> getExampleName ex </> escaped (escapeExperiment e) <.> "svg"
-             | e <- experiments
-             , ex <- examples
-           ]
-        ++ [ build </> getExampleName ex </> T.unpack (humanName ver) </> escaped (escapeExperiment e) <.> mode <.> "svg"
-             | e <- experiments,
-               ex <- examples,
-               ver <- versions,
-               mode <- ["", "diff"]
-           ]
-
-buildRules :: FilePattern -> MkBuildRules BuildSystem example -> Rules ()
-buildRules build MkBuildRules{..} = do
-  build -/- "*/commitid" %> \out -> do
-      alwaysRerun
-
-      let [_,ver,_] = splitDirectories out
-      mbEntry <- find ((== T.pack ver) . humanName) <$> askOracle (GetVersions ())
-      let gitThing :: String
-          gitThing = maybe ver (T.unpack . gitName) mbEntry
-      Stdout commitid <- command [] "git" ["rev-list", "-n", "1", gitThing]
-      writeFileChanged out $ init commitid
-
-  priority 10 $ [ build -/- "HEAD/ghcide"
-                , build -/- "HEAD/ghc.path"
-                ]
-    &%> \[out, ghcpath] -> do
-      liftIO $ createDirectoryIfMissing True $ dropFileName out
-      need =<< getDirectoryFiles "." ["src//*.hs", "exe//*.hs", "ghcide.cabal"]
-      bt <- askOracle $ GetBuildSystem ()
-      cmd_ $ buildProject bt (takeDirectory out)
-      ghcLoc <- liftIO $ findGhc bt "."
-      writeFile' ghcpath ghcLoc
-
-  [ build -/- "*/ghcide",
-    build -/- "*/ghc.path"
-    ]
-    &%> \[out, ghcpath] -> do
-      let [b, ver, _] = splitDirectories out
-      liftIO $ createDirectoryIfMissing True $ dropFileName out
-      commitid <- readFile' $ b </> ver </> "commitid"
-      cmd_ $ "git worktree add bench-temp " ++ commitid
-      buildSystem <- askOracle $ GetBuildSystem ()
-      flip actionFinally (cmd_ (s "git worktree remove bench-temp --force")) $ do
-        ghcLoc <- liftIO $ findGhc buildSystem "bench-temp"
-        cmd_ [Cwd "bench-temp"] $ buildGhcide buildSystem (".." </> takeDirectory out)
-        writeFile' ghcpath ghcLoc
-
-benchRules :: FilePattern -> Resource -> MkBuildRules BuildSystem Example -> Rules ()
-benchRules build benchResource MkBuildRules{benchProject} = do
-  priority 0 $
-    [ build -/- "*/*/*.csv",
-      build -/- "*/*/*.benchmark-gcStats",
-      build -/- "*/*/*.log"
-    ]
-      &%> \[outcsv, _outGc, outLog] -> do
-        let [_, exampleName, ver, exp] = splitDirectories outcsv
-        example <- fromMaybe (error $ "Unknown example " <> exampleName)
-                    <$> askOracle (GetExample exampleName)
-        buildSystem <- askOracle  $ GetBuildSystem ()
-        samples <- askOracle $ GetSamples ()
-        liftIO $ createDirectoryIfMissing True $ dropFileName outcsv
-        let ghcide = build </> ver </> "ghcide"
-            ghcpath = build </> ver </> "ghc.path"
-            experiment = Escaped exp
-        need [ghcide, ghcpath]
-        ghcPath <- readFile' ghcpath
-        withResource benchResource 1 $ do
-          command_
-              [ EchoStdout False,
-                FileStdout outLog,
-                RemEnv "NIX_GHC_LIBDIR",
-                RemEnv "GHC_PACKAGE_PATH",
-                AddPath [takeDirectory ghcPath, "."] []
-              ]
-              "ghcide-bench" $
-              benchProject buildSystem BenchProject{..}
-          cmd_ Shell $ "mv *.benchmark-gcStats " <> dropFileName outcsv
-
-csvRules :: FilePattern -> Rules ()
-csvRules build = do
-  build -/- "*/*/results.csv" %> \out -> do
-      experiments <- askOracle $ GetExperiments ()
-
-      let allResultFiles = [takeDirectory out </> escaped (escapeExperiment e) <.> "csv" | e <- experiments]
-      allResults <- traverse readFileLines allResultFiles
-
-      let header = head $ head allResults
-          results = map tail allResults
-      writeFileChanged out $ unlines $ header : concat results
-
-  build -/- "results.csv" %> \out -> do
-    examples <- map getExampleName <$> askOracle (GetExamples ())
-    let allResultFiles = [build </> e </> "results.csv" | e <- examples]
-
-    allResults <- traverse readFileLines allResultFiles
-
-    let header = head $ head allResults
-        results = map tail allResults
-        header' = "example, " <> header
-        results' = zipWith (\e -> map (\l -> e <> ", " <> l)) examples results
-
-    writeFileChanged out $ unlines $ header' : concat results'
-
-  build -/- "*/results.csv" %> \out -> do
-    versions <- map (T.unpack . humanName) <$> askOracle (GetVersions ())
-    let example = takeFileName $ takeDirectory out
-        allResultFiles =
-          [build </> example </> v </> "results.csv" | v <- versions]
-
-    allResults <- traverse readFileLines allResultFiles
-
-    let header = head $ head allResults
-        results = map tail allResults
-        header' = "version, " <> header
-        results' = zipWith (\v -> map (\l -> v <> ", " <> l)) versions results
-
-    writeFileChanged out $ unlines $ header' : interleave results'
-
-svgRules :: FilePattern -> Rules ()
-svgRules build = do
-  priority 2 $
-    build -/- "*/*/*.diff.svg" %> \out -> do
-      let [b, example, ver, exp_] = splitDirectories out
-          exp = Escaped $ dropExtension $ dropExtension exp_
-      prev <- askOracle $ GetParent $ T.pack ver
-
-      runLog <- loadRunLog b example exp ver
-      runLogPrev <- loadRunLog b example exp $ T.unpack prev
-
-      let diagram = Diagram Live [runLog, runLogPrev] title
-          title = show (unescapeExperiment exp) <> " - live bytes over time compared"
-      plotDiagram True diagram out
-
-
-  priority 1 $
-    build -/- "*/*/*.svg" %> \out -> do
-      let [b, example, ver, exp] = splitDirectories out
-      runLog <- loadRunLog b example (Escaped $ dropExtension exp) ver
-      let diagram = Diagram Live [runLog] title
-          title = ver <> " live bytes over time"
-      plotDiagram True diagram out
-  build -/- "*/*.svg" %> \out -> do
-    let exp = Escaped $ dropExtension $ takeFileName out
-        example = takeFileName $ takeDirectory out
-    versions <- askOracle $ GetVersions ()
-
-    runLogs <- forM (filter include versions) $ \v -> do
-      loadRunLog build example exp $ T.unpack $ humanName v
-
-    let diagram = Diagram Live runLogs title
-        title = show (unescapeExperiment exp) <> " - live bytes over time"
-    plotDiagram False diagram out
-
---------------------------------------------------------------------------------
-
-buildGhcide :: BuildSystem -> String -> String
-buildGhcide Cabal out = unwords
-    ["cabal install"
-    ,"exe:ghcide"
-    ,"--installdir=" ++ out
-    ,"--install-method=copy"
-    ,"--overwrite-policy=always"
-    ,"--ghc-options -rtsopts"
-    ]
-buildGhcide Stack out =
-    "stack --local-bin-path=" <> out
-        <> " build ghcide:ghcide --copy-bins --ghc-options -rtsopts"
-
-
-findGhcDefault :: BuildSystem -> FilePath -> IO FilePath
-findGhcDefault Cabal _cwd =
-    liftIO $ fromMaybe (error "ghc is not in the PATH") <$> findExecutable "ghc"
-findGhcDefault Stack cwd = do
-    Stdout ghcLoc <- cmd [Cwd cwd] (s "stack exec which ghc")
-    return ghcLoc
-
-benchGhcide :: BuildSystem -> BenchProject Example -> [String]
-benchGhcide buildSystem BenchProject{..} =
-  [ "--timeout=3000",
-    "-v",
-    "--samples=" <> show samples,
-    "--csv=" <> outcsv,
-    "--ghcide-options= +RTS -I0.5 -RTS",
-    "--select",
-    unescaped (unescapeExperiment experiment)
-  ] ++
-  exampleToOptions example ++
-  [ "--stack" | Stack == buildSystem]
+ghcideBuildRules :: MkBuildRules BuildSystem
+ghcideBuildRules = MkBuildRules findGhcDefault "ghcide" buildGhcide
 
 --------------------------------------------------------------------------------
 
@@ -357,217 +100,64 @@ data Config buildSystem = Config
   deriving (Generic, Show)
   deriving anyclass (FromJSON)
 
-data GitCommit = GitCommit
-  { -- | A git hash, tag or branch name (e.g. v0.1.0)
-    gitName :: Text,
-    -- | A human understandable name (e.g. fix-collisions-leak)
-    name :: Maybe Text,
-    -- | The human understandable name of the parent, if specified explicitly
-    parent :: Maybe Text,
-    -- | Whether to include this version in the top chart
-    include :: Bool
-  }
-  deriving (Binary, Eq, Hashable, Generic, NFData, Show)
+createBuildSystem :: (Resource -> Rules a) -> Rules a
+createBuildSystem userRules = do
+  readConfig <- newCache $ \fp -> need [fp] >> liftIO (readConfigIO fp)
 
-instance FromJSON GitCommit where
-  parseJSON (String s) = pure $ GitCommit s Nothing Nothing True
-  parseJSON (Object (toList -> [(name, String gitName)])) =
-    pure $ GitCommit gitName (Just name) Nothing True
-  parseJSON (Object (toList -> [(name, Object props)])) =
-    GitCommit
-      <$> props .:? "git"  .!= name
-      <*> pure (Just name)
-      <*> props .:? "parent"
-      <*> props .:? "include" .!= True
-  parseJSON _ = empty
+  _ <- addOracle $ \GetExperiments {} -> experiments <$> readConfig config
+  _ <- addOracle $ \GetVersions {} -> versions <$> readConfig config
+  _ <- addOracle $ \GetExamples{} -> examples <$> readConfig config
+  _ <- addOracle $ \(GetParent name) -> findPrev name . versions <$> readConfig config
+  _ <- addOracle $ \(GetExample name) -> find (\e -> getExampleName e == name) . examples <$> readConfig config
+  _ <- addOracle $ \GetBuildSystem {} -> buildTool <$> readConfig config
 
-instance ToJSON GitCommit where
-  toJSON GitCommit {..} =
-    case name of
-      Nothing -> String gitName
-      Just n -> Object $ fromList [(n, String gitName)]
+  benchResource <- newResource "ghcide-bench" 1
 
-humanName :: GitCommit -> Text
-humanName GitCommit {..} = fromMaybe gitName name
+  userRules benchResource
 
-findPrev :: Text -> [GitCommit] -> Text
-findPrev name (x : y : xx)
-  | humanName y == name = humanName x
-  | otherwise = findPrev name (y : xx)
-findPrev name _ = name
+--------------------------------------------------------------------------------
 
-data BuildSystem = Cabal | Stack
-  deriving (Eq, Read, Show, Generic)
-  deriving (Binary, Hashable, NFData)
-
-instance FromJSON BuildSystem where
-    parseJSON x = fromString . map toLower <$> parseJSON x
-      where
-        fromString "stack" = Stack
-        fromString "cabal" = Cabal
-        fromString other = error $ "Unknown build system: " <> other
-
-instance ToJSON BuildSystem where
-    toJSON = toJSON . show
-----------------------------------------------------------------------------------------------------
-
--- | A line in the output of -S
-data Frame = Frame
-  { allocated, copied, live :: !Int,
-    user, elapsed, totUser, totElapsed :: !Double,
-    generation :: !Int
-  }
-  deriving (Show)
-
-instance Read Frame where
-  readPrec = do
-    spaces
-    allocated <- readPrec @Int <* spaces
-    copied <- readPrec @Int <* spaces
-    live <- readPrec @Int <* spaces
-    user <- readPrec @Double <* spaces
-    elapsed <- readPrec @Double <* spaces
-    totUser <- readPrec @Double <* spaces
-    totElapsed <- readPrec @Double <* spaces
-    _ <- readPrec @Int <* spaces
-    _ <- readPrec @Int <* spaces
-    "(Gen:  " <- replicateM 7 get
-    generation <- readPrec @Int
-    ')' <- get
-    return Frame {..}
-    where
-      spaces = readP_to_Prec $ const P.skipSpaces
-
-data TraceMetric = Allocated | Copied | Live | User | Elapsed
-  deriving (Generic, Enum, Bounded, Read)
-
-instance Show TraceMetric where
-  show Allocated = "Allocated bytes"
-  show Copied = "Copied bytes"
-  show Live = "Live bytes"
-  show User = "User time"
-  show Elapsed = "Elapsed time"
-
-frameMetric :: TraceMetric -> Frame -> Double
-frameMetric Allocated = fromIntegral . allocated
-frameMetric Copied = fromIntegral . copied
-frameMetric Live = fromIntegral . live
-frameMetric Elapsed = elapsed
-frameMetric User = user
-
-data Diagram = Diagram
-  { traceMetric :: TraceMetric,
-    runLogs :: [RunLog],
-    title :: String
-  }
-  deriving (Generic)
-
--- | A file path containing the output of -S for a given run
-data RunLog = RunLog
-  { runVersion :: !String,
-    _runExample :: !String,
-    _runExperiment :: !String,
-    runFrames :: ![Frame],
-    runSuccess :: !Bool
-  }
-
-loadRunLog :: HasCallStack => FilePath -> String -> Escaped FilePath -> FilePath -> Action RunLog
-loadRunLog buildF example exp ver = do
-  let log_fp = buildF </> example </> ver </> escaped exp <.> "benchmark-gcStats"
-      csv_fp = replaceExtension log_fp "csv"
-  log <- readFileLines log_fp
-  csv <- readFileLines csv_fp
-  let frames =
-        [ f
-          | l <- log,
-            Just f <- [readMaybe l],
-            -- filter out gen 0 events as there are too many
-            generation f == 1
+buildGhcide :: BuildSystem -> [CmdOption] -> FilePath -> Action ()
+buildGhcide Cabal args out = do
+    command_ args "cabal"
+        ["install"
+        ,"exe:ghcide"
+        ,"--installdir=" ++ out
+        ,"--install-method=copy"
+        ,"--overwrite-policy=always"
+        ,"--ghc-options=-rtsopts"
         ]
-      success = case map (T.split (== ',') . T.pack) csv of
-          [_header, _name:s:_] | Just s <- readMaybe (T.unpack s) -> s
-          _ -> error $ "Cannot parse: " <> csv_fp
-  return $ RunLog ver example (dropExtension $ escaped exp) frames success
 
-plotDiagram :: Bool -> Diagram -> FilePath -> Action ()
-plotDiagram includeFailed t@Diagram {traceMetric, runLogs} out = do
-  let extract = frameMetric traceMetric
-  liftIO $ E.toFile E.def out $ do
-    E.layout_title .= title t
-    E.setColors myColors
-    forM_ runLogs $ \rl ->
-      when (includeFailed || runSuccess rl) $ E.plot $ do
-        lplot <- E.line
-            (runVersion rl ++ if runSuccess rl then "" else " (FAILED)")
-            [ [ (totElapsed f, extract f)
-                | f <- runFrames rl
-                ]
-            ]
-        return (lplot E.& E.plot_lines_style . E.line_width E.*~ 2)
+buildGhcide Stack args out =
+    command_ args "stack"
+        ["--local-bin-path=" <> out
+        ,"build"
+        ,"ghcide:ghcide"
+        ,"--copy-bins"
+        ,"--ghc-options=-rtsopts"
+        ]
 
-s :: String -> String
-s = id
+findGhcDefault :: BuildSystem -> FilePath -> IO FilePath
+findGhcDefault Cabal _cwd =
+    liftIO $ fromMaybe (error "ghc is not in the PATH") <$> findExecutable "ghc"
+findGhcDefault Stack cwd = do
+    Stdout ghcLoc <- cmd [Cwd cwd] "stack exec which ghc"
+    return ghcLoc
 
-(-/-) :: FilePattern -> FilePattern -> FilePattern
-a -/- b = a <> "/" <> b
+benchGhcide
+  :: Natural -> BuildSystem -> [CmdOption] -> BenchProject Example -> Action ()
+benchGhcide samples buildSystem args BenchProject{..} =
+  command_ args "ghcide-bench" $
+    [ "--timeout=3000",
+        "-v",
+        "--samples=" <> show samples,
+        "--csv="     <> outcsv,
+        "--ghcide="  <> exePath,
+        "--select",
+        unescaped (unescapeExperiment experiment)
+    ] ++
+    exampleToOptions example ++
+    [ "--stack" | Stack == buildSystem
+    ] ++
+    exeExtraArgs
 
-newtype Escaped a = Escaped {escaped :: a}
-
-newtype Unescaped a = Unescaped {unescaped :: a}
-  deriving newtype (Show, FromJSON, ToJSON, Eq, NFData, Binary, Hashable)
-
-escapeExperiment :: Unescaped String -> Escaped String
-escapeExperiment = Escaped . map f . unescaped
-  where
-    f ' ' = '_'
-    f other = other
-
-unescapeExperiment :: Escaped String -> Unescaped String
-unescapeExperiment = Unescaped . map f . escaped
-  where
-    f '_' = ' '
-    f other = other
-
-interleave :: [[a]] -> [a]
-interleave = concat . transpose
-
-myColors :: [E.AlphaColour Double]
-myColors = map E.opaque
-  [ E.blue
-  , E.green
-  , E.red
-  , E.orange
-  , E.yellow
-  , E.violet
-  , E.black
-  , E.gold
-  , E.brown
-  , E.hotpink
-  , E.aliceblue
-  , E.aqua
-  , E.beige
-  , E.bisque
-  , E.blueviolet
-  , E.burlywood
-  , E.cadetblue
-  , E.chartreuse
-  , E.coral
-  , E.crimson
-  , E.darkblue
-  , E.darkgray
-  , E.darkgreen
-  , E.darkkhaki
-  , E.darkmagenta
-  , E.deeppink
-  , E.dodgerblue
-  , E.firebrick
-  , E.forestgreen
-  , E.fuchsia
-  , E.greenyellow
-  , E.lightsalmon
-  , E.seagreen
-  , E.olive
-  , E.sandybrown
-  , E.sienna
-  , E.peru
-  ]
