+++ shake-bench/shake-bench.cabal
@@ -1,0 +1,44 @@
+cabal-version: 2.2
+name:          shake-bench
+version:       0.1.0.0
+synopsis:      Build rules for historical benchmarking
+license:       Apache-2.0
+license-file:  LICENSE
+author:        Pepe Iborra
+maintainer:    pepeiborra@gmail.com
+category:      Development
+build-type:    Simple
+description:
+  A library Shake rules to build and run benchmarks for multiple revisions of a project.
+  An example of usage can be found in the ghcide benchmark suite
+
+library
+  exposed-modules:  Development.Benchmark.Rules
+  hs-source-dirs:   src
+  build-depends:
+        aeson,
+        base == 4.*,
+        Chart,
+        Chart-diagrams,
+        diagrams,
+        diagrams-svg,
+        directory,
+        extra >= 1.7.2,
+        filepath,
+        shake,
+        text
+  default-language: Haskell2010
+  default-extensions:
+        BangPatterns
+        DeriveFunctor
+        DeriveGeneric
+        FlexibleContexts
+        GeneralizedNewtypeDeriving
+        LambdaCase
+        NamedFieldPuns
+        RecordWildCards
+        ScopedTypeVariables
+        StandaloneDeriving
+        TupleSections
+        TypeApplications
+        ViewPatterns
